done all

+++ lab_08/inc/graph.h
@@ -4,12 +4,12 @@
 typedef struct
 {
     int count;
-    double **data;
+    int **data;
 } matrix_t;
 
 matrix_t *create_matrix(int count);
-double **allocate_matrix(int count);
-void free_matrix(double **data, int count);
+int **allocate_matrix(int count);
+void free_matrix(int **data, int count);
 int check_empty_matrix(matrix_t *graph);
 int check_graph_null(matrix_t *graph);
