+++ lab_08/inc/task.h
@@ -4,8 +4,8 @@
 #include "graph.h"
 
 int find_tops_with_Dijkstra(matrix_t *graph, int *count);
-void Dijkstra(matrix_t *graph, double distance[graph->count], int top);
-int get_top_and_way(int *top, double *min_way, matrix_t *graph);
+void Dijkstra(matrix_t *graph, int distance[graph->count], int top);
+int get_top_and_way(int *top, int *min_way, matrix_t *graph);
 void time_get();
 
 #endif // __TASK_H__
